% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_vec.R
\name{as.array,LatentNeuroVec-method}
\alias{as.array,LatentNeuroVec-method}
\title{Convert LatentNeuroVec to Array}
\usage{
\S4method{as.array}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A \code{LatentNeuroVec} object to convert.}

\item{...}{Not used.}
}
\value{
A 4D array representing the full reconstructed data.
}
\description{
Converts a \code{LatentNeuroVec} to a standard R 4D array.
}
\details{
This method reconstructs the full 4D array from the latent representation, which can be 
memory-intensive for large datasets. It calculates:
\deqn{array[,,, t] = Map(basis[t,] \%*\% t(loadings)) + offset}
for each time point, where values outside the mask are set to zero.
}
