% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_clustered_vec.R
\name{as.data.frame,H5ReducedClusteredVec-method}
\alias{as.data.frame,H5ReducedClusteredVec-method}
\title{Convert an H5ReducedClusteredVec to a data.frame}
\usage{
\S4method{as.data.frame}{H5ReducedClusteredVec}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An \code{H5ReducedClusteredVec} instance.}

\item{row.names}{Optional row names (usually \code{NULL}).}

\item{optional}{Passed to \code{\link{as.data.frame}} internals.}

\item{...}{Additional arguments for \code{as.data.frame}.}
}
\value{
A \code{data.frame} with \code{nTime} rows and \code{nClusters} columns.
}
\description{
Calls \code{as.matrix(x)} and then converts to \code{data.frame}.
}
