% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_clustered_vec.R
\name{as.data.frame,H5ReducedClusteredVecSeq-method}
\alias{as.data.frame,H5ReducedClusteredVecSeq-method}
\title{Convert an H5ReducedClusteredVecSeq to data.frames}
\usage{
\S4method{as.data.frame}{H5ReducedClusteredVecSeq}(x, bind_scans = FALSE, ...)
}
\arguments{
\item{x}{An \code{H5ReducedClusteredVecSeq} object.}

\item{bind_scans}{\code{logical} indicating whether to combine all scans into
one big data.frame. If \code{FALSE}, returns a list of data.frames, one per scan.}

\item{...}{Further arguments passed to \code{as.data.frame()}.}
}
\value{
A \code{list} of data.frames if \code{bind_scans=FALSE}, or a single
  combined data.frame if \code{bind_scans=TRUE}.
}
\description{
Similar to \code{as.matrix} for \code{H5ReducedClusteredVecSeq}, but each matrix
is turned into a \code{data.frame}. If \code{bind_scans=TRUE}, row-binds them into
one combined \code{data.frame} (adding a \code{ScanName} column for identification).
}
