% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cluster_array.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,H5ParcellatedScanSummary-method}
\title{Convert to Data Frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{H5ParcellatedScanSummary}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object for which a `as.data.frame` method is defined.}

\item{row.names}{A character vector giving the row names for the data frame, or \code{NULL}.}

\item{optional}{Logical. If \code{TRUE}, setting row names and converting column names 
is optional.}

\item{...}{Additional arguments passed to the underlying as.data.frame methods.}
}
\description{
These methods convert `fmristore` specific objects to data frames,
implementing the S4 generic function \code{\link[base]{as.data.frame}} from the `base` package.
}
\seealso{
Other H5Cluster: 
\code{\link{H5ParcellatedMultiScan-class}},
\code{\link{[,H5ParcellatedScan,ANY,ANY,ANY-method}},
\code{\link{[,H5ParcellatedScan,ANY,missing,ANY-method}},
\code{\link{as.matrix}()},
\code{\link{dim}()},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{series}()},
\code{\link{show,H5ParcellatedScan-method}}
}
\concept{H5Cluster}
