% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_clustered_vec.R
\name{as.matrix,H5ReducedClusteredVec-method}
\alias{as.matrix,H5ReducedClusteredVec-method}
\title{Convert an H5ReducedClusteredVec to a matrix}
\usage{
\S4method{as.matrix}{H5ReducedClusteredVec}(x)
}
\arguments{
\item{x}{An \code{H5ReducedClusteredVec} instance.}
}
\value{
A numeric matrix of dimension \code{[nTime, nClusters]}.
}
\description{
Reads the summary dataset \code{/scans/<scan_name>/clusters_summary/summary_data}
from HDF5, returning it as a numeric matrix \code{[nTime, nClusters]}. If
\code{cluster_names} are available, they become the column names.
}
