% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_clustered_vec.R
\name{as.matrix,H5ReducedClusteredVecSeq-method}
\alias{as.matrix,H5ReducedClusteredVecSeq-method}
\title{Convert an H5ReducedClusteredVecSeq to matrices}
\usage{
\S4method{as.matrix}{H5ReducedClusteredVecSeq}(x, bind_scans = FALSE)
}
\arguments{
\item{x}{An \code{H5ReducedClusteredVecSeq} object (multiple scans).}

\item{bind_scans}{\code{logical}. If \code{FALSE}, returns a \code{list} of matrices,
one per scan. If \code{TRUE} and all matrices have the same number of columns
(clusters), row-binds them into one large matrix.}
}
\value{
A \code{list} of matrices (if \code{bind_scans=FALSE}) or a single matrix
  (if \code{bind_scans=TRUE}).
}
\description{
For each scan in the sequence, constructs an \code{H5ReducedClusteredVec} on the fly,
reads its summary dataset, and returns the result(s).
}
