% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cluster_array.R
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,H5ParcellatedScanSummary-method}
\title{Convert to Matrix}
\usage{
as.matrix(x, ...)

\S4method{as.matrix}{H5ParcellatedScanSummary}(x)
}
\arguments{
\item{x}{An object for which a `as.matrix` method is defined.}

\item{...}{Additional arguments passed to specific as.matrix methods.}
}
\description{
These methods convert `fmristore` specific objects to matrices,
implementing the S4 generic function \code{\link[base]{as.matrix}} from the `base` package.
}
\seealso{
Other H5Cluster: 
\code{\link{H5ParcellatedMultiScan-class}},
\code{\link{[,H5ParcellatedScan,ANY,ANY,ANY-method}},
\code{\link{[,H5ParcellatedScan,ANY,missing,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{dim}()},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{series}()},
\code{\link{show,H5ParcellatedScan-method}}
}
\concept{H5Cluster}
