% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R
\name{$,H5ClusteredExperiment-method}
\alias{$,H5ClusteredExperiment-method}
\title{Accessor Methods for H5ClusteredExperiment
Access Slots/Properties using `$`}
\usage{
\S4method{$}{H5ClusteredExperiment}(x, name)
}
\arguments{
\item{x}{An `H5ClusteredExperiment` object.}

\item{name}{The name of the property or slot to access (`mask`, `clusters`, `obj`, `runs`, `scan_metadata`, `cluster_metadata`).}
}
\value{
The requested object or value.
}
\description{
Provides convenient access to shared properties like `mask`, `clusters`,
and the underlying HDF5 file object (`obj`) by retrieving them from the
first run object stored within the experiment.
Also provides access to the experiment's own slots (`runs`, `scan_metadata`, `cluster_metadata`).
}
\seealso{
Other H5Clustered: 
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusteredExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series # Keep @name if you want the generic itself to be findable by ?series}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}
}
\concept{H5Clustered}
