% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R
\name{$,H5ParcellatedMultiScan-method}
\alias{$,H5ParcellatedMultiScan-method}
\title{Accessor Methods for H5ParcellatedMultiScan
Access Slots/Properties using `$`}
\usage{
\S4method{$}{H5ParcellatedMultiScan}(x, name)
}
\arguments{
\item{x}{An `H5ParcellatedMultiScan` object.}

\item{name}{The name of the property or slot to access (`mask`, `clusters`, `obj`, `runs`, `scan_metadata`, `cluster_metadata`).}
}
\value{
The requested object or value.
}
\description{
Provides convenient access to shared properties like `mask`, `clusters`,
and the underlying HDF5 file object (`obj`) by retrieving them from the
first run object stored within the experiment.
Also provides access to the experiment's own slots (`runs`, `scan_metadata`, `cluster_metadata`).
}
\seealso{
Other H5Parcellated: 
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{h5file,H5ParcellatedArray-method}},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}
}
\concept{H5Parcellated}
