% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cluster_metadata}
\alias{cluster_metadata}
\title{Get cluster metadata}
\usage{
cluster_metadata(x)
}
\arguments{
\item{x}{The object from which to retrieve the cluster metadata}
}
\value{
A data frame or other structure containing metadata for each cluster
}
\description{
This generic returns any available metadata associated with clusters in an object.
}
