% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_vec.R
\name{concat,LatentNeuroVec,LatentNeuroVec-method}
\alias{concat,LatentNeuroVec,LatentNeuroVec-method}
\title{Concatenate LatentNeuroVec Objects}
\usage{
\S4method{concat}{LatentNeuroVec,LatentNeuroVec}(x, y, ...)
}
\arguments{
\item{x}{First \code{LatentNeuroVec}.}

\item{y}{Second \code{LatentNeuroVec}.}

\item{...}{Additional \code{LatentNeuroVec} objects to concatenate.}
}
\value{
A \code{\link[neuroim2]{NeuroVecSeq-class}} containing all input vectors.
}
\description{
Concatenates two or more \code{LatentNeuroVec} objects along the temporal dimension.
}
\details{
1. Checks for compatibility
2. Creates a sequence container
3. Preserves individual object properties
}
\examples{
\dontrun{
# Concatenate multiple LatentNeuroVec objects
combined <- concat(lvec1, lvec2, lvec3)
dim(combined)
}

}
\seealso{
\code{\link[neuroim2]{NeuroVecSeq-class}}
}
