% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5neurovec.R
\name{DenseNeuroVec,H5NeuroVec}
\alias{DenseNeuroVec,H5NeuroVec}
\title{Convert DenseNeuroVec to H5NeuroVec}
\arguments{
\item{from}{A \code{DenseNeuroVec} object.}
}
\value{
An \code{H5NeuroVec} referencing the newly created HDF5 file.
}
\description{
Converts a \code{DenseNeuroVec} object to an \code{H5NeuroVec} by writing it to HDF5.
}
\keyword{internal}
