% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol_hdf5.R
\name{DenseNeuroVol,H5NeuroVol}
\alias{DenseNeuroVol,H5NeuroVol}
\title{Convert DenseNeuroVol to H5NeuroVol}
\arguments{
\item{from}{A \code{DenseNeuroVol} object.}
}
\value{
An \code{H5NeuroVol} referencing the resulting HDF5 file.
}
\description{
Converts a \code{DenseNeuroVol} to \code{H5NeuroVol} by writing it to disk in HDF5 format.
}
\keyword{internal}
