% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_methods.R
\name{detect_h5_type}
\alias{detect_h5_type}
\title{Detect the type of data in an HDF5 file}
\usage{
detect_h5_type(h5obj)
}
\arguments{
\item{h5obj}{An H5File object or file path to an HDF5 file}
}
\value{
Character string indicating the type, one of:
  - "H5NeuroVol" - 3D brain volume
  - "H5NeuroVec" - 4D brain time series
  - "H5NeuroVecSeq" - Sequence of 4D scans
  - "H5ParcellatedMultiScan" - Multi-run parcellated experiment
  - "H5ParcellatedScan" - Single parcellated run (full data)
  - "H5ParcellatedScanSummary" - Single parcellated run (summary)
  - "LatentNeuroVec" - Latent representation
  - "LabeledVolumeSet" - Labeled brain regions
  - "unknown" - Type could not be determined
}
\description{
Examines the structure and attributes of an HDF5 file to determine
what type of fmristore object it contains.
}
\examples{
\dontrun{
# Detect type from file path
type <- detect_h5_type("data.h5")

# Detect type from open H5File
h5 <- H5File$new("data.h5", "r")
type <- detect_h5_type(h5)
h5$close_all()
}
}
