% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cluster_array.R
\name{dim}
\alias{dim}
\alias{dim,H5ParcellatedScan-method}
\title{Get Dimensions of an Object (Methods for base R Generic)}
\usage{
\S4method{dim}{H5ParcellatedScan}(x)
}
\arguments{
\item{x}{An object for which a `dim` method is defined.}
}
\value{
A numeric vector of dimensions. The length and interpretation of the
  vector depend on the specific class of `x`.
}
\description{
These methods retrieve the dimensions of `fmristore` specific objects,
implementing the S4 generic function \code{\link[base]{dim}} from the `base` package.

For objects that represent neuroimaging data, this typically returns a numeric
vector indicating the size of each dimension (e.g., X, Y, Z, Time).
}
\seealso{
\code{base::\link[base]{dim}}, specific methods like
  \code{dim,H5ParcellatedScan-method} (if defined and exported).

Other H5Cluster: 
\code{\link{H5ParcellatedMultiScan-class}},
\code{\link{[,H5ParcellatedScan,ANY,ANY,ANY-method}},
\code{\link{[,H5ParcellatedScan,ANY,missing,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{series}()},
\code{\link{show,H5ParcellatedScan-method}}
}
\concept{H5Cluster}
\keyword{internal}
