% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_methods.R
\name{exists_in_h5}
\alias{exists_in_h5}
\title{Check if a path exists in an HDF5 file}
\usage{
exists_in_h5(h5, path)
}
\arguments{
\item{h5}{An H5File object}

\item{path}{Character string path to check}
}
\value{
Logical indicating if path exists
}
\description{
Safely checks if a path exists in an HDF5 file
}
\keyword{internal}
