% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_methods.R
\name{extract_cluster_data}
\alias{extract_cluster_data}
\title{Extract Full Cluster Data}
\usage{
extract_cluster_data(nvec, clusters)
}
\arguments{
\item{nvec}{A NeuroVec object containing the voxel data}

\item{clusters}{A ClusteredNeuroVol object defining the clusters}
}
\value{
A list with one matrix per cluster (voxels x time)
}
\description{
Extract voxel-level data organized by clusters
}
\keyword{internal}
