% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_hdf5.R
\name{guess_h5_type}
\alias{guess_h5_type}
\title{Internal helper to guess the HDF5 type from an R vector}
\usage{
guess_h5_type(x)
}
\description{
Internal helper to guess the HDF5 type from an R vector
}
\keyword{internal}
