% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_methods.R
\name{h5_attr_exists}
\alias{h5_attr_exists}
\title{Check if an attribute exists in an HDF5 object}
\usage{
h5_attr_exists(h5, attr_name)
}
\arguments{
\item{h5}{An H5File or H5Group object}

\item{attr_name}{Character string attribute name}
}
\value{
Logical indicating if attribute exists
}
\description{
Safely checks if an attribute exists
}
\keyword{internal}
