% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_methods.R
\name{h5_attr_safe}
\alias{h5_attr_safe}
\title{Get HDF5 attribute safely}
\usage{
h5_attr_safe(h5, attr_name)
}
\arguments{
\item{h5}{An H5File or H5Group object}

\item{attr_name}{Character string attribute name}
}
\value{
Attribute value or NULL
}
\description{
Gets an attribute value or returns NULL if not found
}
\keyword{internal}
