% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H5_ATTRS}
\alias{H5_ATTRS}
\title{HDF5 Attribute Constants}
\format{
A named list of character strings.
}
\usage{
H5_ATTRS
}
\value{
A named list with HDF5 attribute name constants.
}
\description{
Named constants for HDF5 attribute names used throughout the package.
}
\examples{
H5_ATTRS$N_TIME
}
\keyword{datasets}
