% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_utils.R
\name{h5_dataset_dims}
\alias{h5_dataset_dims}
\title{Obtain the dimensions of an HDF5 dataset}
\usage{
h5_dataset_dims(h5, path)
}
\arguments{
\item{h5}{An open `H5File` handle.}

\item{path}{Path to the dataset within the file.}
}
\value{
Integer vector of dataset dimensions.
}
\description{
Obtain the dimensions of an HDF5 dataset
}
\keyword{internal}
