% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H5_DSETS}
\alias{H5_DSETS}
\title{HDF5 Dataset Name Constants}
\format{
A named list of character strings.
}
\usage{
H5_DSETS
}
\value{
A named list with HDF5 dataset name constants.
}
\description{
Named constants for HDF5 dataset names used throughout the package.
}
\examples{
H5_DSETS$DATA
}
\keyword{datasets}
