% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H5_PATHS}
\alias{H5_PATHS}
\title{HDF5 Path Constants}
\format{
An object of class \code{list} of length 21.
}
\usage{
H5_PATHS
}
\description{
Central definition of HDF5 file layout constants used throughout the package.
This provides a single source of truth for the HDF5 file structure.
}
\examples{
# Access constants
H5_PATHS$SCANS_GROUP  # "/scans"
sprintf(H5_PATHS$CLUSTER_DSET_TPL, "run1", 5)  # "/scans/run1/clusters/cluster_5"
}
\keyword{datasets}
