% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5ClusteredVec-class}
\alias{H5ClusteredVec-class}
\title{H5ClusteredVec Class}
\description{
A class representing a single clustered time series scan backed by HDF5.
Provides a NeuroVec interface while maintaining HDF5 backing.
}
\section{Slots}{

\describe{
\item{\code{obj}}{The H5File object}

\item{\code{scan_name}}{Name of this scan}

\item{\code{mask}}{LogicalNeuroVol defining the brain mask}

\item{\code{clusters}}{ClusteredNeuroVol containing cluster assignments}
}}

