% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5ClusteredVecSeq-class}
\alias{H5ClusteredVecSeq-class}
\title{H5ClusteredVecSeq Class}
\description{
A class representing a sequence of clustered time series backed by an HDF5 file.
Provides a NeuroVecSeq-like interface while maintaining HDF5 backing for efficient
data access.
}
\section{Slots}{

\describe{
\item{\code{obj}}{The H5File object}

\item{\code{scan_names}}{Names of available scans}

\item{\code{mask}}{LogicalNeuroVol defining the brain mask}

\item{\code{clusters}}{ClusteredNeuroVol containing cluster assignments}

\item{\code{scan_metadata}}{List of metadata for each scan}

\item{\code{cluster_metadata}}{Data frame of cluster metadata}
}}

