% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_vec.R
\name{H5ClusteredVecSeq}
\alias{H5ClusteredVecSeq}
\title{Create an H5ClusteredVecSeq object}
\usage{
H5ClusteredVecSeq(
  obj,
  scan_names,
  mask,
  clusters,
  scan_metadata,
  cluster_metadata
)
}
\arguments{
\item{obj}{An \code{\link[hdf5r]{H5File}} object (open for reading).}

\item{scan_names}{A \code{character} vector of scan IDs to include.}

\item{mask}{A \code{\link[neuroim2]{LogicalNeuroVol}} giving the 3D mask.}

\item{clusters}{A \code{\link[neuroim2]{ClusteredNeuroVol}} describing cluster assignments.}

\item{scan_metadata}{A \code{list} of metadata for each scan.}

\item{cluster_metadata}{A \code{data.frame} of cluster metadata.}
}
\value{
A new \code{H5ClusteredVecSeq} instance.
}
\description{
Constructs a multi-scan \code{H5ClusteredVecSeq} object referencing a set of
clustered time-series datasets stored in an HDF5 file.
}
