% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_array.R, R/cluster_experiment.R,
%   R/h5neurovecseq.R
\name{h5file,H5ParcellatedArray-method}
\alias{h5file,H5ParcellatedArray-method}
\alias{h5file,H5ParcellatedMultiScan-method}
\alias{h5file,H5ParcellatedScan-method}
\alias{h5file,H5ParcellatedScanSummary-method}
\alias{h5file,H5NeuroVecSeq-method}
\title{Get the HDF5 file object via generic}
\usage{
\S4method{h5file}{H5ParcellatedArray}(x)

\S4method{h5file}{H5ParcellatedMultiScan}(x)

\S4method{h5file}{H5ParcellatedScan}(x)

\S4method{h5file}{H5ParcellatedScanSummary}(x)

\S4method{h5file}{H5NeuroVecSeq}(x)
}
\arguments{
\item{x}{H5ParcellatedScanSummary object}
}
\value{
The HDF5 file object from the first run.

The H5File object handle

The H5File object handle
}
\description{
Get the HDF5 file object via generic

Get the H5File object for H5ParcellatedScan objects

Get the H5File object for H5ParcellatedScanSummary objects
}
\seealso{
Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}

Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}

Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}
}
\concept{H5Parcellated}
