% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec_hdf5.R
\name{H5NeuroVec}
\alias{H5NeuroVec}
\title{H5NeuroVec Constructor}
\usage{
H5NeuroVec(file_name)
}
\arguments{
\item{file_name}{The name of the 4-dimensional image file in HDF5 format.}
}
\value{
An instance of the \code{\link{H5NeuroVec-class}} class.
}
\description{
This function constructs an H5NeuroVec object, which represents a four-dimensional
brain image stored in the HDF5 format.
}
\seealso{
\code{\link[neuroim2]{NeuroVec-class}} for the base 4D brain image class.
}
