% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5NeuroVecSeq-class}
\alias{H5NeuroVecSeq-class}
\title{H5NeuroVecSeq Class}
\description{
Represents a collection of 4D scans stored in a single HDF5 file.
Each scan is accessed lazily via an internal \code{H5NeuroVec}.
}
\section{Slots}{

\describe{
\item{\code{obj}}{An \code{H5File} handle to the underlying HDF5 file.}

\item{\code{vecs}}{A list of \code{H5NeuroVec} objects, one per scan.}
}}

