% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol_hdf5.R
\name{H5NeuroVol}
\alias{H5NeuroVol}
\title{H5NeuroVol Constructor}
\usage{
H5NeuroVol(file_name)
}
\arguments{
\item{file_name}{A \code{character} string giving the path to a 3D HDF5 file.}
}
\value{
A new \code{\link{H5NeuroVol-class}} instance.
}
\description{
Constructs an \code{\link{H5NeuroVol}} object representing a 3D brain volume
stored in HDF5 format.
}
\seealso{
\code{\link[neuroim2]{NeuroVol-class}} for the base 3D brain volume class.
}
