% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5ParcellatedMultiScan-class}
\alias{H5ParcellatedMultiScan-class}
\title{H5ParcellatedMultiScan Class}
\description{
Represents a collection of clustered neuroimaging runs (scans) stored within a single HDF5 file.
It acts as a container for `H5ParcellatedScan` and/or `H5ParcellatedScanSummary` objects,
along with associated metadata.

This class facilitates managing multiple runs that share the same mask and cluster definitions.
}
\section{Slots}{

\describe{
\item{\code{runs}}{A `list` where each element is an object inheriting from `H5ParcellatedArray`
(typically `H5ParcellatedScan` or `H5ParcellatedScanSummary`).}

\item{\code{scan_metadata}}{A `list` containing metadata for each scan in the `runs` list.}

\item{\code{cluster_metadata}}{A `data.frame` containing metadata associated with the clusters.}
}}

\seealso{
\code{\link{H5ParcellatedScan-class}}, \code{\link{H5ParcellatedScanSummary-class}}

Other H5Cluster: 
\code{\link{[,H5ParcellatedScan,ANY,ANY,ANY-method}},
\code{\link{[,H5ParcellatedScan,ANY,missing,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{dim}()},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{series}()},
\code{\link{show,H5ParcellatedScan-method}}
}
\concept{H5Cluster}
