% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_h5_helpers.R
\name{is_h5file}
\alias{is_h5file}
\title{Check if an object is an H5File handle}
\usage{
is_h5file(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical TRUE if inherits from H5File, FALSE otherwise.
}
\description{
Check if an object is an H5File handle
}
\keyword{internal}
