% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_vec.R
\name{linear_access,H5ClusteredVecSeq,numeric-method}
\alias{linear_access,H5ClusteredVecSeq,numeric-method}
\title{linear_access for H5ClusteredVecSeq}
\usage{
\S4method{linear_access}{H5ClusteredVecSeq,numeric}(x, i)
}
\arguments{
\item{x}{An \code{H5ClusteredVecSeq} object.}

\item{i}{A numeric vector of mask-based voxel indices \code{(1..sum(x@mask))}.}
}
\value{
A numeric matrix of size \code{[length(i), sum_of_all_scan_timepoints]}.
}
\description{
Returns the time series for a set of mask-based voxel indices across all
scans in the sequence, concatenated along the time dimension.

If the sequence has scans S1, S2, ..., each with T1, T2, ... time points,
the total time dimension T = T1 + T2 + ... .
}
