% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec_hdf5.R
\name{linear_access,H5NeuroVec,numeric-method}
\alias{linear_access,H5NeuroVec,numeric-method}
\title{A bounding-box approach for linear indexing of a 4D H5NeuroVec}
\usage{
\S4method{linear_access}{H5NeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{A \code{H5NeuroVec} object (4D).}

\item{i}{Numeric vector of linear indices into the 4D array.}
}
\value{
A numeric vector of length \code{length(i)}, in the same order as \code{i}.
}
\description{
Converts linear indices \code{i} into 4D subscripts \code{(x,y,z,t)} in
\code{dim(x)}, then reads the minimal bounding region from HDF5.
}
