% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol_hdf5.R
\name{linear_access,H5NeuroVol,numeric-method}
\alias{linear_access,H5NeuroVol,numeric-method}
\title{A simple bounding-box approach for linear indexing of H5NeuroVol}
\usage{
\S4method{linear_access}{H5NeuroVol,numeric}(x, i)
}
\arguments{
\item{x}{An \code{\link{H5NeuroVol}} object.}

\item{i}{A numeric vector of linear voxel indices (1-based).}
}
\value{
A numeric vector of length \code{length(i)} containing the voxel values
  in the same order as \code{i}.
}
\description{
Given a set of linear voxel indices in a 3D volume, this method maps them to
\code{(x,y,z)} coordinates, reads the minimal bounding box from HDF5 once,
and extracts the requested values in the correct order.
}
