% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_volume.R
\name{linear_access,LabeledVolumeSet,numeric-method}
\alias{linear_access,LabeledVolumeSet,numeric-method}
\title{linear_access for LabeledVolumeSet}
\usage{
\S4method{linear_access}{LabeledVolumeSet,numeric}(x, i)
}
\arguments{
\item{x}{A \code{LabeledVolumeSet} object.}

\item{i}{numeric vector of 1D indices in the full [X * Y * Z * nVols] space.}
}
\value{
A numeric vector of length \code{length(i)}, in the same order as \code{i}.
}
\description{
linear_access for LabeledVolumeSet
}
