% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_vec.R
\name{loadings}
\alias{loadings}
\alias{loadings,LatentNeuroVec-method}
\title{Get the loadings matrix (spatial components)}
\usage{
loadings(x, ...)

\S4method{loadings}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{An object, likely a LatentNeuroVec or similar}

\item{...}{Additional arguments}
}
\value{
The loadings matrix (typically voxels x components)
}
\description{
Get the loadings matrix (spatial components)
}
\examples{
\dontrun{
# For LatentNeuroVec:
if (!is.null(fmristore:::create_minimal_LatentNeuroVec)) {
  lnv <- NULL
  tryCatch({
    # Helper creates a mask, n_mask_voxels determined internally or by arg
    lnv <- fmristore:::create_minimal_LatentNeuroVec(
      space_dims = c(4L, 4L, 2L),
      n_time = 10L,
      n_comp = 3L
    )
    l_matrix <- loadings(lnv)
    # Dimensions should be n_voxels_in_mask x n_comp
    print(dim(l_matrix))
  }, error = function(e) {
    message("loadings example for LatentNeuroVec failed: ", e$message)
  })
} else {
  message("Skipping loadings example for LatentNeuroVec: helper not available.")
}
}

}
