% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_array.R, R/cluster_experiment.R,
%   R/latent_vec.R
\name{mask,H5ParcellatedArray-method}
\alias{mask,H5ParcellatedArray-method}
\alias{mask,H5ParcellatedMultiScan-method}
\alias{mask,LatentNeuroVec-method}
\title{Get the mask object via generic}
\usage{
\S4method{mask}{H5ParcellatedArray}(x)

\S4method{mask}{H5ParcellatedMultiScan}(x)

\S4method{mask}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{H5ParcellatedMultiScan object}
}
\value{
The mask object from the first run.
}
\description{
Get the mask object via generic
}
\seealso{
Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{h5file,H5ParcellatedArray-method}},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}
}
\concept{H5Parcellated}
