% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R
\name{n_scans,H5ClusterExperiment-method}
\alias{n_scans,H5ClusterExperiment-method}
\title{Get number of scans}
\usage{
\S4method{n_scans}{H5ClusterExperiment}(x)
}
\arguments{
\item{x}{H5ClusterExperiment object}
}
\value{
Integer number of scans.
}
\description{
Get number of scans
}
\seealso{
Other H5Cluster: 
\code{\link{$,H5ClusterExperiment-method}},
\code{\link{H5ClusterExperiment}},
\code{\link{H5ClusterExperiment-class}},
\code{\link{H5ClusterRun-class}},
\code{\link{H5ClusterRunSummary-class}},
\code{\link{H5ClusteredArray-class}},
\code{\link{[,H5ClusterRun,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusterExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{scan_metadata,H5ClusterExperiment-method}},
\code{\link{scan_names,H5ClusterExperiment-method}},
\code{\link{series}()},
\code{\link{series_concat}()},
\code{\link{show,H5ClusterRun-method}}
}
\concept{H5Cluster}
