% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{n_scans}
\alias{n_scans}
\title{Get the number of scans}
\usage{
n_scans(x)
}
\arguments{
\item{x}{The object from which to retrieve the number of scans}
}
\value{
An integer representing the number of scans
}
\description{
This generic returns the number of scans in an object (e.g. a sequence of fMRI scans).
}
