% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/labeled_vec.R
\name{names}
\alias{names}
\alias{names,LabeledVolumeSet-method}
\title{Get Object Names (Methods for base R Generic)}
\usage{
\S4method{names}{LabeledVolumeSet}(x)
}
\arguments{
\item{x}{An object for which a `names` method is defined.}
}
\value{
A character vector of names associated with the object.
}
\description{
These methods retrieve the names from `fmristore` specific objects,
implementing the S4 generic function \code{\link[base]{names}} from the `base` package.
}
\seealso{
\code{base::\link[base]{names}}
}
\keyword{internal}
