% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_hdf5.R
\name{print.LabeledVolumeSet}
\alias{print.LabeledVolumeSet}
\title{Print method for LabeledVolumeSet}
\usage{
print.LabeledVolumeSet(x, ...)
}
\arguments{
\item{x}{A \code{\link[neuroim2]{LabeledVolumeSet}} object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for LabeledVolumeSet
}
