% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_methods.R
\name{read_dataset_typed}
\alias{read_dataset_typed}
\title{Read Dataset Method Implementation}
\usage{
read_dataset_typed(file, type, ...)
}
\arguments{
\item{file}{Path to HDF5 file}

\item{type}{Character string indicating type}

\item{...}{Additional arguments passed to constructors}
}
\value{
The appropriate fmristore object
}
\description{
Internal function to read based on known type
}
\keyword{internal}
