% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_h5_helpers.R
\name{read_h5_clusters_to_ClusteredNeuroVol}
\alias{read_h5_clusters_to_ClusteredNeuroVol}
\title{Read /cluster_map (+ coords) into ClusteredNeuroVol}
\usage{
read_h5_clusters_to_ClusteredNeuroVol(h5, mask, map_dset = "/cluster_map")
}
\description{
Read /cluster_map (+ coords) into ClusteredNeuroVol
}
\keyword{internal}
