% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_h5_helpers.R
\name{read_h5_mask_to_LogicalNeuroVol}
\alias{read_h5_mask_to_LogicalNeuroVol}
\title{Read /mask into a LogicalNeuroVol}
\usage{
read_h5_mask_to_LogicalNeuroVol(h5, dset = "/mask", ref_space)
}
\arguments{
\item{h5}{An open `H5File` handle.}

\item{dset}{The path to the mask dataset (default: "/mask").}

\item{ref_space}{A `NeuroSpace` object representing the expected space of the mask.}
}
\value{
A `LogicalNeuroVol` object corresponding to the mask dataset.
}
\description{
Read /mask into a LogicalNeuroVol
}
\keyword{internal}
