% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_utils.R
\name{safe_h5_close}
\alias{safe_h5_close}
\title{Safely close an HDF5 file handle}
\usage{
safe_h5_close(h5)
}
\arguments{
\item{h5}{An object, expected to be an H5File handle.}
}
\description{
Checks if the object is a valid H5File handle before attempting
to close it, silencing potential errors from double-closing etc.
}
\keyword{internal}
