% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scan_metadata}
\alias{scan_metadata}
\title{Get scan metadata}
\usage{
scan_metadata(x)
}
\arguments{
\item{x}{The object from which to retrieve the scan metadata}
}
\value{
A list (or other structure) containing metadata for each scan
}
\description{
This generic returns any available metadata associated with each scan.
}
