% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scan_names}
\alias{scan_names}
\title{Get the scan names}
\usage{
scan_names(x)
}
\arguments{
\item{x}{The object from which to retrieve the scan names}
}
\value{
A character vector of scan names
}
\description{
This generic returns a character vector of scan names or labels.
}
