% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cluster_array.R,
%   R/h5neurovec.R, R/latent_vec.R
\name{series # Keep @name if you want the generic itself to be findable by ?series}
\alias{series # Keep @name if you want the generic itself to be findable by ?series}
\alias{series}
\alias{series,H5ClusteredRunFull,numeric-method}
\alias{series,H5ClusteredRunFull,matrix-method}
\alias{series,H5ClusteredRunSummary,ANY-method}
\alias{series,H5NeuroVec,integer-method}
\alias{series,H5NeuroVec,numeric-method}
\alias{series,H5NeuroVec,matrix-method}
\alias{series,LatentNeuroVec,integer-method}
\alias{series,LatentNeuroVec,numeric-method}
\title{Extract Time Series Data}
\usage{
series(x, i, ...)

\S4method{series}{H5ClusteredRunFull,numeric}(x, i, j, k, ...)

\S4method{series}{H5ClusteredRunFull,matrix}(x, i, ...)

\S4method{series}{H5ClusteredRunSummary,ANY}(x, i, ...)

\S4method{series}{H5NeuroVec,integer}(x, i, j, k, drop = TRUE)

\S4method{series}{H5NeuroVec,numeric}(x, i, j, k, drop = TRUE)

\S4method{series}{H5NeuroVec,matrix}(x, i)

\S4method{series}{H5NeuroVec,integer}(x, i, j, k, drop = TRUE)

\S4method{series}{H5NeuroVec,numeric}(x, i, j, k, drop = TRUE)

\S4method{series}{LatentNeuroVec,integer}(x, i, j, k, drop = TRUE)

\S4method{series}{LatentNeuroVec,numeric}(x, i, j, k, drop = TRUE)
}
\arguments{
\item{x}{The object from which to extract series data.}

\item{i}{An index, typically specifying voxels or elements.
The interpretation of `i` depends on the specific method.
It can be numeric indices, a matrix of coordinates, etc.}

\item{...}{Additional arguments passed to specific methods (e.g., `drop`).}

\item{j}{Optional y-coordinate or further index specification used by some methods.}

\item{k}{Optional z-coordinate or further index specification used by some methods.}

\item{drop}{Whether to drop dimensions of length 1}
}
\value{
A matrix or vector containing the time series data. The exact format
  (e.g., time x voxels, or a simple vector for a single voxel) depends on the
  method and arguments.
}
\description{
A generic function to extract time series data from an object.
Specific methods will determine how data is extracted based on the object's class
and the indices provided.
}
\examples{
# This is a generic function; examples are provided with its specific methods.
# For example, see help("series,H5ClusteredRunFull-method")

}
\seealso{
Methods for this generic are available for classes like
  \code{\link{H5ClusteredRunFull}}.

Other H5Clustered: 
\code{\link{$,H5ClusteredExperiment-method}},
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusteredExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}

Other H5Clustered: 
\code{\link{$,H5ClusteredExperiment-method}},
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusteredExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}

Other H5Clustered: 
\code{\link{$,H5ClusteredExperiment-method}},
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusteredExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}
}
\concept{H5Clustered}
