% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_vec.R
\name{show,H5ClusteredVec-method}
\alias{show,H5ClusteredVec-method}
\title{Pretty Printer for H5ClusteredVec}
\usage{
\S4method{show}{H5ClusteredVec}(object)
}
\description{
Displays a concise, nicely formatted summary of an \code{H5ClusteredVec},
including the scan name, the cluster assignments, and the underlying
HDF5 file information.
}
