% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_vec.R
\name{show,H5ClusteredVecSeq-method}
\alias{show,H5ClusteredVecSeq-method}
\title{Pretty Printer for H5ClusteredVecSeq}
\usage{
\S4method{show}{H5ClusteredVecSeq}(object)
}
\description{
Displays a concise, nicely formatted summary of an \code{H5ClusteredVecSeq},
including the number of scans, the cluster assignments, and the underlying
HDF5 file information.
}
