% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec_hdf5.R
\name{show,H5NeuroVec-method}
\alias{show,H5NeuroVec-method}
\title{Display an H5NeuroVec object}
\usage{
\S4method{show}{H5NeuroVec}(object)
}
\arguments{
\item{object}{An \code{H5NeuroVec} object to show.}
}
\description{
Prints a concise summary of the H5NeuroVec object, including
4D dimensions, spacing, origin, and orientation.
}
