% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5neurovol.R
\name{show,H5NeuroVol-method}
\alias{show,H5NeuroVol-method}
\title{Show Method for H5NeuroVol}
\usage{
\S4method{show}{H5NeuroVol}(object)
}
\arguments{
\item{object}{The H5NeuroVol object to display.}
}
\description{
Displays a summary of the H5NeuroVol object, including dimensions,
spacing, origin, and HDF5 file information, without reading voxel data.
}
