% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_clustered_vec.R
\name{show,H5ReducedClusteredVecSeq-method}
\alias{show,H5ReducedClusteredVecSeq-method}
\title{Pretty Printer for H5ReducedClusteredVecSeq}
\usage{
\S4method{show}{H5ReducedClusteredVecSeq}(object)
}
\description{
Displays a concise but nicely formatted summary of an \code{H5ReducedClusteredVecSeq},
including information about the scans, clusters, and data location.
}
