% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_volume.R
\name{show,LabeledVolumeSet-method}
\alias{show,LabeledVolumeSet-method}
\title{show method for LabeledVolumeSet}
\usage{
\S4method{show}{LabeledVolumeSet}(object)
}
\arguments{
\item{object}{A \code{\link{LabeledVolumeSet}} instance}
}
\description{
Displays essential info about a \code{LabeledVolumeSet}, including spatial dims,
number of volumes, label previews, spacing, origin, orientation (if known),
and storage paths.
}
