% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_vec.R
\name{show,LatentNeuroVec-method}
\alias{show,LatentNeuroVec-method}
\title{Show Method for LatentNeuroVec Objects}
\usage{
\S4method{show}{LatentNeuroVec}(object)
}
\arguments{
\item{object}{A \code{LatentNeuroVec} to display.}
}
\description{
Displays a formatted summary of a \code{LatentNeuroVec} with colored output.
}
\details{
Uses the \pkg{crayon} package for colored terminal output, showing:
\itemize{
  \item Dimensions & components
  \item Memory usage
  \item Sparsity info
  \item Space metadata
}
}
\keyword{internal}
