% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_vec.R
\name{[,H5ClusteredVecSeq,numeric,numeric,ANY-method}
\alias{[,H5ClusteredVecSeq,numeric,numeric,ANY-method}
\title{4D Subsetting of H5ClusteredVecSeq}
\usage{
\S4method{[}{H5ClusteredVecSeq,numeric,numeric,ANY}(x, i, j, k, l, ..., drop = TRUE)
}
\arguments{
\item{x}{An \code{H5ClusteredVecSeq} object.}

\item{i, j, k}{Numeric vectors specifying 3D spatial indexing (if missing, full range).
If only \code{i} is provided, it may be interpreted as mask-based indices if
\code{j,k} are missing.}

\item{l}{Numeric vector for the global time dimension across the entire sequence.
E.g., if the first scan has 200 timepoints, the second 180, then times 1..200
map to scan 1, times 201..380 map to scan 2, etc.}

\item{...}{Not used.}

\item{drop}{Logical, if \code{TRUE}, drop single-length dimensions in the result.}
}
\value{
A 4D array of shape \code{[length(i), length(j), length(k), length(l)]}
  if fully specified, or fewer dims if partial. If only a single i or a single l
  is requested, dimensions may be dropped if \code{drop=TRUE}.
}
\description{
Interprets the combined time dimension of all scans in the sequence as one
concatenated axis. Then extracts (i,j,k) in space plus t in time. If t spans
multiple scans, partial subsets are read from each \code{H5ClusteredVec} object
and concatenated appropriately.
}
