% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec_hdf5.R
\name{[,H5NeuroVec,numeric,numeric,ANY-method}
\alias{[,H5NeuroVec,numeric,numeric,ANY-method}
\title{Subset a 4D H5NeuroVec with arbitrary numeric i, j, k, l (bounding box)}
\usage{
\S4method{[}{H5NeuroVec,numeric,numeric,ANY}(x, i, j, k, l, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{H5NeuroVec} (4D).}

\item{i, j, k, l}{Numeric vectors of indices (possibly non-contiguous).}

\item{...}{Ignored.}

\item{drop}{Logical; drop dimensions of size 1?}
}
\value{
An array of dimension \code{length(i) × length(j) × length(k) × length(l)},
  dropped if \code{drop=TRUE}.
}
\description{
Handles \code{x[i,j,k,l]} subsetting. We read a bounding hyper-slab from HDF5 once,
then pick out the requested points in correct order, avoiding loops in R.
}
