% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol_hdf5.R
\name{[,H5NeuroVol,ANY,ANY,ANY-method}
\alias{[,H5NeuroVol,ANY,ANY,ANY-method}
\title{3D bracket subsetting for H5NeuroVol (handles partial arguments)}
\usage{
\S4method{[}{H5NeuroVol,ANY,ANY,ANY}(x, i, j, k, ..., drop = TRUE)
}
\arguments{
\item{x}{An \code{H5NeuroVol} instance}

\item{i, j, k}{Numeric (or integer) index vectors for each dimension. If missing,
we take the full range in that dimension.}

\item{...}{Unused}

\item{drop}{Logical: whether to drop dimensions of size 1. Default \code{TRUE}.}
}
\value{
A numeric \code{array} of shape \code{c(length(i), length(j), length(k))},
  or fewer dims if \code{drop=TRUE}.
}
\description{
Allows \code{h5vol[i, j, k]} where each of \code{i,j,k} may be missing.
Missing arguments default to the entire range in that dimension.
Zero-length arguments immediately yield an empty array of the correct shape.
}
