% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_volume.R
\name{[,LabeledVolumeSet,numeric,numeric,ANY-method}
\alias{[,LabeledVolumeSet,numeric,numeric,ANY-method}
\title{4D Array-like subsetting for LabeledVolumeSet}
\usage{
\S4method{[}{LabeledVolumeSet,numeric,numeric,ANY}(x, i, j, k, l, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{LabeledVolumeSet} object.}

\item{i}{Numeric indices for the 1st dimension (x).}

\item{j}{Numeric indices for the 2nd dimension (y).}

\item{k}{Numeric indices for the 3rd dimension (z).}

\item{l}{Numeric indices for the 4th dimension (label).}

\item{...}{Ignored.}

\item{drop}{Logical, whether to drop singleton dimensions.}
}
\value{
An R array after subsetting, or a lower-dimensional array if \code{drop=TRUE}.
}
\description{
4D Array-like subsetting for LabeledVolumeSet
}
