% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_vec.R
\name{[,LatentNeuroVec,numeric,numeric,ANY-method}
\alias{[,LatentNeuroVec,numeric,numeric,ANY-method}
\title{4D subsetting for LatentNeuroVec}
\usage{
\S4method{[}{LatentNeuroVec,numeric,numeric,ANY}(x, i, j, k, l, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{LatentNeuroVec} object}

\item{i, j, k}{Numeric index vectors for the 3D spatial dims. If missing,
all voxels are included.}

\item{l}{Numeric index vector (timepoints). If missing, all time.}

\item{...}{Not used}

\item{drop}{\code{logical} drop dims of size 1?}
}
\value{
An array with shape \code{[length(i), length(j), length(k), length(l)]}
  or, if \code{i,j,k} are missing, \code{[dim(x)[1], dim(x)[2], dim(x)[3], length(l)]}.
  Then we reconstruct each timepoint on-the-fly using \code{basis[l, ]} and
  \code{loadings}, plus \code{offset}.
}
\description{
Allows \code{latent_vec[i, j, k, l]} style subsetting:
\itemize{
  \item \code{i,j,k} are currently assumed missing or full range
  \item \code{l} can be a numeric vector (subset of timepoints)
}
}
