% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cluster_array.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,H5ClusteredRunSummary-method}
\title{Convert to Data Frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{H5ClusteredRunSummary}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object for which a `as.data.frame` method is defined.}
}
\description{
These methods convert `fmristore` specific objects to data frames,
implementing the S4 generic function \code{\link[base]{as.data.frame}} from the `base` package.
}
\seealso{
Other H5Clustered: 
\code{\link{$,H5ClusteredExperiment-method}},
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.matrix}()},
\code{\link{cluster_metadata,H5ClusteredExperiment-method}},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series # Keep @name if you want the generic itself to be findable by ?series}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}
}
\concept{H5Clustered}
