% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_h5_helpers.R
\name{close_h5_safely}
\alias{close_h5_safely}
\title{Safely close an HDF5 object (file, group, dataset, attribute, dataspace, property list)}
\usage{
close_h5_safely(h5obj)
}
\description{
Checks if the object is valid and not NULL before attempting to close.
}
\keyword{internal}
