% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R
\name{cluster_metadata,H5ParcellatedMultiScan-method}
\alias{cluster_metadata,H5ParcellatedMultiScan-method}
\title{Get cluster metadata}
\usage{
\S4method{cluster_metadata}{H5ParcellatedMultiScan}(x)
}
\arguments{
\item{x}{H5ParcellatedMultiScan object}
}
\value{
Data frame of cluster metadata.
}
\description{
Get cluster metadata
}
\seealso{
Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{clusters}()},
\code{\link{h5file,H5ParcellatedArray-method}},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ParcellatedMultiScan-method}},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}
}
\concept{H5Parcellated}
