% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R
\name{cluster_metadata,H5ClusteredExperiment-method}
\alias{cluster_metadata,H5ClusteredExperiment-method}
\title{Get cluster metadata}
\usage{
\S4method{cluster_metadata}{H5ClusteredExperiment}(x)
}
\arguments{
\item{x}{H5ClusteredExperiment object}
}
\value{
Data frame of cluster metadata.
}
\description{
Get cluster metadata
}
\seealso{
Other H5Clustered: 
\code{\link{$,H5ClusteredExperiment-method}},
\code{\link{H5ClusteredArray-class}},
\code{\link{H5ClusteredExperiment}},
\code{\link{H5ClusteredExperiment-class}},
\code{\link{H5ClusteredRunFull-class}},
\code{\link{H5ClusteredRunSummary-class}},
\code{\link{[,H5ClusteredRunFull,ANY,ANY,ANY-method}},
\code{\link{as.data.frame}()},
\code{\link{as.matrix}()},
\code{\link{clusters}()},
\code{\link{dim}()},
\code{\link{h5file,H5ClusteredArray-method}},
\code{\link{linear_access-methods}},
\code{\link{make_run_full}()},
\code{\link{make_run_summary}()},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{n_scans,H5ClusteredExperiment-method}},
\code{\link{scan_metadata,H5ClusteredExperiment-method}},
\code{\link{scan_names,H5ClusteredExperiment-method}},
\code{\link{series # Keep @name if you want the generic itself to be findable by ?series}},
\code{\link{series_concat}()},
\code{\link{show,H5ClusteredRunFull-method}}
}
\concept{H5Clustered}
