% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_vec.R
\name{concat,LatentNeuroVec,LatentNeuroVec-method}
\alias{concat,LatentNeuroVec,LatentNeuroVec-method}
\title{Concatenate LatentNeuroVec Objects}
\usage{
\S4method{concat}{LatentNeuroVec,LatentNeuroVec}(x, y, ...)
}
\arguments{
\item{x}{First \code{\link[neuroim2]{LatentNeuroVec-class}} object}

\item{y}{Second \code{\link[neuroim2]{LatentNeuroVec-class}} object}

\item{...}{Additional LatentNeuroVec objects to concatenate}
}
\value{
A \code{\link[neuroim2]{NeuroVecSeq-class}} object containing all input vectors
}
\description{
Concatenates two or more LatentNeuroVec objects along the temporal dimension.
}
\details{
The method:
\itemize{
  \item Validates compatibility of input objects
  \item Creates a sequence container
  \item Preserves individual object properties
}
}
\examples{
\dontrun{
# Create two LatentNeuroVec objects
lvec1 <- LatentNeuroVec(basis1, loadings1, space1, mask)
lvec2 <- LatentNeuroVec(basis2, loadings2, space2, mask)

# Concatenate them
combined <- concat(lvec1, lvec2)

# Access as a sequence
dim(combined)  # Shows combined dimensions
combined[[1]]  # First volume
}

}
\seealso{
\code{\link[neuroim2]{NeuroVecSeq-class}} for the return type
}
