% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H5_ATTRS}
\alias{H5_ATTRS}
\title{HDF5 attribute constants}
\format{
An object of class \code{list} of length 7.
}
\usage{
H5_ATTRS
}
\description{
HDF5 attribute constants
}
\keyword{datasets}
