% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H5_DSETS}
\alias{H5_DSETS}
\title{HDF5 dataset name constants}
\format{
An object of class \code{list} of length 3.
}
\usage{
H5_DSETS
}
\description{
HDF5 dataset name constants
}
\keyword{datasets}
