% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec_hdf5.R
\name{H5NeuroVec}
\alias{H5NeuroVec}
\title{H5NeuroVec Constructor}
\usage{
H5NeuroVec(file_name)
}
\arguments{
\item{file_name}{The path to a 4D HDF5 file.}
}
\value{
A new \code{\link{H5NeuroVec-class}} instance.
}
\description{
Constructs an \code{\link{H5NeuroVec}} object, which represents a 4D brain image
stored in HDF5 format.
}
\seealso{
\code{\link[neuroim2]{NeuroVec-class}} for details on 4D brain images.
}
