% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5neurovecseq.R
\name{H5NeuroVecSeq}
\alias{H5NeuroVecSeq}
\title{H5NeuroVecSeq constructor}
\usage{
H5NeuroVecSeq(file)
}
\arguments{
\item{file}{Path to the HDF5 file created by `neurovecseq_to_h5()`.}
}
\value{
An object of class `H5NeuroVecSeq`.
}
\description{
Load a NeuroVecSeq stored in HDF5 format. Each scan is stored
under `/scans/<name>/data` and will be returned as an `H5NeuroVec`.
}
