% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol_hdf5.R
\name{H5NeuroVol}
\alias{H5NeuroVol}
\title{H5NeuroVol Constructor}
\usage{
H5NeuroVol(file_name)
}
\arguments{
\item{file_name}{The name of the 3-dimensional image file in HDF5 format.}
}
\value{
An instance of the \code{\link{H5NeuroVol-class}} class.
}
\description{
This function constructs an H5NeuroVol object, which represents a three-dimensional
brain image stored in the HDF5 format.
}
\seealso{
\code{\link[neuroim2]{NeuroVol-class}} for the base 3D brain volume class.
}
