% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_experiment.R, R/h5neurovecseq.R
\name{n_scans,H5ParcellatedMultiScan-method}
\alias{n_scans,H5ParcellatedMultiScan-method}
\alias{n_scans,H5NeuroVecSeq-method}
\title{Get number of scans}
\usage{
\S4method{n_scans}{H5ParcellatedMultiScan}(x)

\S4method{n_scans}{H5NeuroVecSeq}(x)
}
\arguments{
\item{x}{H5ParcellatedMultiScan object}
}
\value{
Integer number of scans.
}
\description{
Get number of scans
}
\seealso{
Other H5Parcellated: 
\code{\link{$,H5ParcellatedMultiScan-method}},
\code{\link{H5ParcellatedArray-class}},
\code{\link{H5ParcellatedMultiScan}},
\code{\link{H5ParcellatedScan-class}},
\code{\link{H5ParcellatedScanSummary-class}},
\code{\link{close}()},
\code{\link{cluster_metadata,H5ParcellatedMultiScan-method}},
\code{\link{clusters}()},
\code{\link{h5file,H5ParcellatedArray-method}},
\code{\link{mask}()},
\code{\link{matrix_concat}()},
\code{\link{scan_metadata,H5ParcellatedMultiScan-method}},
\code{\link{scan_names,H5ParcellatedMultiScan-method}},
\code{\link{series_concat}()}
}
\concept{H5Parcellated}
