% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/read_dataset_methods.R
\name{read_dataset}
\alias{read_dataset}
\alias{read_dataset,H5File-method}
\title{Read Dataset from HDF5}
\usage{
read_dataset(x, ...)

\S4method{read_dataset}{H5File}(x, type = NULL, ...)
}
\arguments{
\item{x}{The file path or H5File object to read from}

\item{...}{Additional arguments passed to methods}
}
\value{
The appropriate fmristore object
}
\description{
Generic function for reading neuroimaging datasets from HDF5 format
}
